requirements.txt
# only need the git version if there's a bug fixed or a feature in master
#git+https://github.com/plone/plone.server@master#plone.server

plone.server
